+++ lib/fortnite_api_io.dart
@@ -2,3 +2,4 @@ export "src/fortnite_api_io/fortnite_api_io.dart";
 
 export "src/fortnite_api_io/structures/challenges.dart";
 export "src/fortnite_api_io/structures/items.dart";
+export "src/fortnite_api_io/structures/item_sets.dart";